updating a group now brings you back to the group page and updates the group info on your device
updating the search screen is a bit trickier. i thought we'd have to pass in a callback from searchscreen (the top), down to the creatinggroups screen so after updating, searchscreen can navigate to the updated group. instead we can go from creatinggroups to searchscreen and pass a parameter that searchscreen can use to check which item was updated, and load that updated item instead of the older version of the item.

The next problem was loading the new parameter into the searchscreen and grouppostsscreen. seems like you can spawn multiple screens at once so that's what I did.