Majority of displaying friends/requests complete
code to display friends is there, but we need to still test as accept/reject aren't full functional

code for requests is almost done. Need to figure out accept and reject. I think we are missing an id field in the friend object. Without this id field I dont think we can use the api?