we need a query that will list all friendships where either user1 or user2 matches the given user's id could we use key? could we create a new query in the schema? would we need to modify the resolver? modifying the resolver would be easiest imo honestly
EDIT: Nvm a simple filter works! It isn't the most efficient but we can deal with that later. By the time real users have enough friends to fill 1 mb we'll be making money.
Had to include another check for the nextToken being null in the data fetching functions because otherwise the data would repeat itself when the screen rerenders