Incorporated API graphqlOperation into the PaginatedList, renamed it to APIList for clarity. This list will handle fetching data from the backend, and can refresh data and load more data. Its "fetchDataAsync" function can also be used in a ref to grab the results immediately afterwards if you don't want to wait for the state to refresh.
initial and additional amount are built into the apilist now. currently they can't be set from outside, but we should give it that capability later on.
a large amount of code is removed from feedscreen in exchange for 2 extra arguments to APIList. I'd say that's a fair trade.
collectFriends and collectFriendships have been removed from FriendScreen.js. They're called in a few places after manipulating the list, but they can easily be replaced with local changes. If the user wants to refresh the list in case there's discrepency between the local and remote copy, the apilist will handle it.
the APIList can function as a flatlist and a sectionlist (sectionlist ability untested). Since the results from the list's fetchDataAsync function need to be formatted immediately afterwards, we've set the search screen to use a ref. The use case for the search screen is trickier but I think it will generally make things much cleaner once it's been thoroughly bugtested.