fixed a bevy of issues:
in the listgroupitem, matchingname wasn't defined in the props (idk how it could access it then but whatevs) and when it's undefined (as in the mygroups screen) it didn't have a default color causing the app to freeze, now it does
removed the references to "didUserPost" in the subscription events since we arent using that variable anymore
whenever new posts appear, it didnt check whether the new post actually belonged to the feed you're looking at, resulting in new posts from other groups appearing in the feed youre looking at. thats fixed.
the "posts" variable in the subscription event is outdated since the async function referring to it was called when the app starts. similar to the searchscreen, a ref variable fixed this issue (since the ref variable refreshes every frame rather than every re-render?)