Search screen seems mostly stable now! And the all tab works too! The issue with not all results loading before was caused by us intercepting the results in fetchDataAsync in the APIList, preventing it from calling the setDataFunction itself. Meaning onRefresh doesn't work, and loadMore doesn't work, since they both rely on being able to call setDataFunction. The benefit of the approach was that it prevented the list from loading when no query was inputted, but I didn't foresee those other issues. So first I included a new prop in APIList that determines whether or not the list will call fetchDataAsync when it initially loads. It's set to true for the searchscreen, and false by default. Then I included another optional parameter for the fetchDataAsync function (it's the best I could think of atm) which is a function that's called after the fetching is done. If it returns true it will ignore the results, if not it'll load the results like normal. The corresponding changes have been made to the searchscreen when it calls fetchDataAsync.
To get the "all" tab functionality to work was a bit hacky. I learned that sometimes you have to take the messy, hacky solution because it's a lot faster than restructuring the entire screen to add the new functionality in a clean way. So we're sacrificing readability/cleanliness of the code to get this functionality to work faster. And I realized it's a skill in an of itself, "BS'ing" code. Sometimes you HAVE To hack together something quickly and ignore clean coding practices make sure it works to meet a deadline, and doing that in the most efficient way possible is something that can be trained and practiced. Like AP Art History where you have to pull random knowledge out of your brain and put it on the page in a comprehensible way as fast as possible. Idk why it reminds me of that but it does. "BS'ing" essays, "BS'ing" code, same deal. The art of BS.
I tried to make the new changes as small as possible. Two extra state functions are added, one to cover the case where the user's typing a query on the all tab, and the other to cover the case where the user switches from a different tab to the all tab without changing the query. "isAll" covers the first case and "isGroupsFetched" covers the second.
When typing a query on the all tab we need to fetch both userresults and groupresults. One method, the slow but cleaner method, would be to restructure the screen to resemble the friends page, with two separate lists and two separate refs to call two separate fetchdataasync functions. The method I ended up using was to call "fetchDataAsync" twice by switching the type between "all" and "groups". When fetchDataAsync is called in the "all" tab it'll fetch the userResults, since thats the default. Then it'll switch to the groups tab to fetch the groupResults, then switch back to the all tab to append the groupResults to the end. To mask the change, the isAll variable is set to true and is checked accordingly in the styling of the screen.
For switching tabs another variable is used, isGroupsFetched, to prevent an infinite loop in the useeffect for the type variable. Like before, it switches to groups and then switches back to all. Only if the groupresults haven't been fetched yet (aka if groupresults is empty)
Changed double equals to triple requals signs for the miniscule performance boost
Switched the user and group check in the filter prop for the APIList to work with the hacky method I'm using for the all tab
There may be some weirdness since the variables update asynchronously. We'll see if that affects anything.
I just added a check I thought would help prevent unnecessary loads in the useeffect for query and it caused duplicates in the group search screen. Why??? We'll get it tomorrow.
Just realized refreshing and loadingmore wouldn't work on the all tab. *facepalm*

Oh also an important thing, the renderitem checks whether the item is a group or user now by checking if the userID field exists rather than checking the current "type" variable. This allows both group items and user items to be shown in one screen.