The search screen seems to work as well as it did before now! The only issue that sometimes when the searchbar is empty (after deleting everything) a list of results will still show. We attempted to remedy this by adding another check to the voidreturnfunction to see if the query is empty after fetchdataasync executes, but that didn't seem to help. We'll see if the team can find any other bugs later. I also added something to the filter to check if the user's last name starts with the given query. Inspired by Adam appearing under "S"
A new variable was added to apilist which is set to true whenever fetchdataasync runs. This displays the initial loading icon, unless the list is refreshing or loading more items.
nextTok was removed from fetchdataasync, turns out we didnt need it. Instead a do-while loop is used in fetchdataasync to continuously fetch results if the necessary amount isn't fully loaded.
initialAmount and additionalAmount have also been variable-ized to reduce redundancy
a "wasBeginning" variable is made to keep track of the original state of the "beginning" variable argument. the "beginning" argument is changed after fetching so the nextToken doesn't keep staying at null after fetching, resulting in endless fetching of results.
this.setState is run at the end of fetchdataasync, and it's programmed not to rely on any state variables in the middle of execution to prevent asynchronous call weirdness.