friendrequestsbyreceiver now orders friend requests based on when they were made, so the newest ones stay on top modified the accordion to accept an "open" prop, and included a useEffect that allows the parent component to open/close the accordion using this "open" prop. added padding between accordion header and body
included initialloadfunction prop in the apilist to run when the apilist fetches for the first time. was also going to include a "loadinitial" function so the parent component can call the first load itself, but that wouldnt make too much sense and we probably dont need it for this case
layoutanimation added to the apilist so fetching items is animated (untested, try with and without to see difference)
added gray dividers above and below the friendlistitem and friendrequestlistitem. we tested shadows but it didn't look good, and inset shadows don't exist yet.
postitem sets the liked variable before doing the graphql operation, again for responsiveness sake
textlayoutstyle needs to be set to {flex: 1} to avoid overflow, however it can't be set in the profile screen or else it'll push all the other contents off screen. we can probably fix lookupuser to accomodate this, but it isn't necessary yet.
for now, we're hardcoding the path names in playsound since require doesn't allow variable path names. we'll have to adjust this later once we have more sfx.
customsidebarmenu includes layoutanimations for the subscription events, and includes an ondeletefriendship subscription to avoid zaniness when someone unfriends someone else and they can still message them.
instead of playing the celebrate sound whenever friendrequestlist changes (since it'll also play when the component loads) it'll just play in the removefriendrequestlistitem function after the user responds.
checknewrequests called from apilist now using the new initialloadfunction prop
user's profile name at the top of the sidebar is now bold with a white background. to mimic facebook's high contrast ui design.
friendlist header is crossed out when the user has no friends.