APIList is now a purecomponent like the flatlist it extends from onendreachedthreshold is now adjustable allows for getitemlayout prop
postitem is memoized to prevent rerenders
postitem passes the post's preloaded userinfo to profileimageandname

profileimageandname is memoized for possible performance improvements when scrolling across the feed very quickly
removed unused state variable
it accepts a default/preloaded userinfo object now
cache fetch doesnt run if theres preloaded userinfo. addusertocache just runs if there was an issue preloading info from cache.
combined the component that renders when userinfo is null/loading into the base component

feedscreen fetches userinfo of posts from the cache before displaying them to reduce rerenders (as opposed to creating the post objects and having them all read from the cache and rerender asynchronously), plus it looks nicer not having the images load in your face
feedscreen uses usecallback when rendering postitems so when new postitems are added, the older ones dont re-render
increased initialamount and additionalamount and increased onendreachedthreshold so newer posts have time to load before you reach the next batch of posts.
also either making the apilist pure or putting the postitem in a usecallback prevents the items from rerendering when typing on the textinput at the top of the feed. didnt check to see which fixed it.